Created sort function for Leaderboard
MMALeaderboard will no longer be used. All logic is handled inside the Leaderboard component.
This is mainly due to the fact that MMALeaderboard was used to fetch individual users. However, with the sort function, we need all user information first. So it makes more sense to fetch all the user data inside Leaderboard and then calculate/sort their stats.