Updated form validation for AddPick.jsx
Previous form validation logic checked if inputs are valid when the "submit" button was clicked. The handleSubmit function would then run afterwards. It happened to work out by coincidence, but is likely not good practice. Instead, I added useEffects to update the isValid variable. When the "submit" button is clicked, it will check if all isValid values are true, if not, it will update the formError variable to alert the user of errors.